pyproject.toml
[project]
name = "auth"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "python-keycloak>=5.3.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
debug = [
    "debugpy>=1.8.13",
]
